requirements.txt
asgiref==3.7.2
Django==5.0.3
django-bootstrap3==23.6
django-filter==24.1
djangorestframework==3.15.0
sqlparse==0.4.4
typing_extensions==4.10.0
